feat: display balances
